Refactor logic to be usable with other file selectors
